Added touch events in order to support mobile phones
Very simple support for touch events so Remark can be used on mobile
phones. There is a slew of unanswered questions, including how pinching
and double-clicking to zoom works. This patch only solves the very, very
basic problem of getting to the previous (touching the left half of the
slide) and next slide (toucing the right half of the slide).

(Only tested on iPhone)