Use the significantly faster marked instead of Showdown
Finally there is a kickass Markdown parser in JavaScript! This is
significantly faster than Showdown, has far better possibilities of
extensions, and has a strong focus on being consistent with markdown.pl

The reason I've included it far earlier than Showdown, is because
`module.exports` is used in the code, which is a problem with code that
can be used in both Node.js and JavaScript, as `module` is used to
create Node.js modules.