adds a "title" property to slides that is based either on slide number or name
in many cases it might be helpful for the user to be able to look up the name or number of the currently visible slide.  to facilitate this, i added to the <div class='slide'> element a title property.  This is set either to the slide's name using "name:" or the slide's number in the stack.  This can be accessed in the slidein callback using slide.title.