Modified the way slides are hidden, allowing for animated transitions.
This change adds a 'remark-fading' class to a slide when it stops being
visible. This class keeps the old slide visible for 1000ms, on a lowered
z-index, so that new slides can use an animated transition when they appear.
By keeping the old content on the screen, it will look like the new slide is
sliding in/out over the old content.

Adding animations is also fairly easy; If you add
(animate.css)[https://github.com/daneden/animate.css] to your
stylesheet for your presentation, and add a class definition to your
slide markdown:

    class: animated slideInRight

transitions to that slide will be animated.

If you don't use animated transitions, this change has no effect - the
faded slide will be visually "behind" the current slide, making it invisible.