Introduce 3 types of boilerplate HTML files, to get a new person started.
This commit is primarily to reduce barrier to entry; just ask people to
start off with one of these boilerplate HTML files, and they should be
content with the results.

boilerplate-single.html (generated from a template) is a standalone file
that can be used on an offline computer, and without a need of any other
file, like remark.js, etc.

boilerplate-local.html can be used by people who want to present without
an internet connection. This requires at least the accompanying
remark.js file for the presentations to work.

boilerplate-remote.html is for presenting in environments where internet
connectivity is available, and no other accompanying file is needed for
the presntation to work.

Note: I am not including the generated files in the commit, to keep the
patch small.