#85,#97: No longer operate on escaped HTML.
Before, remark would read the source out of the source textarea element,
which would give out escaped HTML. The markdown conversion would eventually
run on the escaped HTML, enabling it to convert recognized markdown entities
nested in HTML, which is by default not possible for all HTML elements using
the 'marked' library.

However, as some markdown translations, like auto-linking URLs, are implemented
by checking if a URL is already inside an <a> tag or not, operating on escaped
HTML made it impossible for the translation to work correctly.

remark now operates on actual HTML. As 'marked' does, per now, not translate
markdown appearing inside tags such as <div>, block content classes, that is
.[...] where ... contains one or more new lines, which is converted into
<div>s, would no longer have their markdown content processed correctly.
To still enable markdown inside block content classes, the parsing process
now no longer serves the content for a slide as a string, but as a list of
content, containing strings for non-content-class stuff, and objects for
content classes, enabling the markdown conversion process to translate the
contents of content classes before they are converted into the corresponding
<span> or <div> tags.

Having 'marked' operate on actual HTML also introduces some annoying bugs,
so for now, it is running in 'pedantic' mode, which makes it conform to the
original markdown as much as possible. This is good as it fixes the bugs I've
seen, and bad as it may introduce other markdown flaws that 'marked' has tried
to fix. Hopefully the 'pedantic' mode can be turned back off some time soon.

Hopefully there aren't to many unintended effects caused by this rather
fundamental change.