Add an option to exclude presenter notes
This PR adds a configuration setting that allows you to exclude presenter notes from all slides. For example:

```js
var slideshow = remark.create({
	includePresenterNotes: false,
});
```

This feature is useful if you only want to publish the slides, without the notes. By default, notes will be include if the options is not present or explicitly set to true.