Move from mocha-phantomjs to mocha-chrome for tests
The PhantomJS and mocha-phantomjs projects have been deprecated and the mocha-chrome is a suggested project to migrate to. I wasn't able to get the mocha-phantomjs to work locally, and the CI was failing due to phantomjs as well. Still need to test that this works on Windows (works fine on MacOS).